% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gsvmod.R
\name{Gsvmod}
\alias{Gsvmod}
\title{G SVM models.}
\usage{
Gsvmod(dat.train, lab.train, dat.test, lab.test)
}
\arguments{
\item{dat.train}{Train data with G features and (k-1)*S/k samples. Parameter k comes from cross-validation scheme and is specified by user (default is 2).}

\item{lab.train}{Class labels for train data.}

\item{dat.test}{Test data with G features and S/k samples.}

\item{lab.test}{Class labels for test data.}
}
\value{
Accuracy scores for models. Each model represents one gene.
}
\description{
Returns accuracy performance of all genes. G support vector machine (SVM) classifiers trained using G different data matrixes, are used to predict labels in test data. Models are ranked based on prediction performances.
}

