% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malanidata.R
\docType{data}
\name{malanidata}
\alias{malanidata}
\title{Dataset for malani package}
\format{A matrix of numeric values, 100 genes, 20 samples and class labels.}
\usage{
malanidata
}
\description{
A numeric matrix G*S contains gene expressions data.
G are the genes (rows) and S are the samples (columns).
}
\examples{
data(malanidata)
}
\keyword{datasets}

