% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairmod.R
\name{pairmod}
\alias{pairmod}
\title{Find best performing pairs}
\usage{
pairmod(X, LX, theta, npair = 10)
}
\arguments{
\item{X}{Original gene expression data matrix. With G rows (number of genes) and S column (number of samples).}

\item{LX}{Class labels.}

\item{theta}{Initial gene set.}

\item{npair}{Given a gene in initial set, top \code{npair} best performing pairs correspond to that gene are selected (Default is 10).}
}
\value{
Best (\code{npair*G/20}) performing pairs.
}
\description{
Combine each gene in initial set with all genes in the original set. Top \code{npair}  pairs are selected to construct the \code{Q} matrix.
}

